fix broken solver
Spreading the "absent" letter evalutations incorrectly accounted for
the "absent" letter evaluations that were added by solver for each "present"
letter evaluation. Current approach is to rather insert an "absent-deduced"
letter evaluation that does not "polute" the actual "absent" letter evalutations.
This may not be preferable, but does fix the solver.